+++ CHANGELOG.md
@@ -174,7 +174,8 @@
 
 
 
-
+* APPSEC
+  * Fixed a bug where updating `notification_emails` along with other activation fields caused a `notification_emails not found` error in the `akamai_appsec_activations` resource.